Handle code blocks
CodeBlocks do not have content.  So, we need to handle the case when the
attribute is in a `<div>` outside of the code block, and when the
attribute is inside the fenced code block.  To place the environment
inside the attribute list of the fenced code block, you must enable the
`+fenced_code_attributes` extension.