Ensure we have a Div
The attribute needs to be on the Div and not the CodeBlock.  Further, we
can generalize more by making sure the content of the Div has content
before we simply insert the RawInline.  If it does not, we need to
replace the element with a Div with a RawBlock before and after the
content.